F&G_index_projet_Barrau_Dalsace_Denis_Walckhoff/finnhub_data/volatilite/requirements.txt
google-cloud-storage
pandas
robotframework-csvlibrary
regex
robotframework-jsonlibrary
numpy